+++ crates/sevenmark-parser/src/parser/bracket/bracket_media.rs
@@ -1,10 +1,10 @@
-use winnow::ascii::multispace0;
 use crate::ast::{Location, MediaElement, SevenMarkElement};
 use crate::parser::ParserInput;
 use crate::parser::element::element_parser;
 use crate::parser::parameter::parameter_core_parser;
 use crate::parser::utils::with_depth_and_trim;
 use winnow::Result;
+use winnow::ascii::multispace0;
 use winnow::combinator::{delimited, opt};
 use winnow::prelude::*;
 use winnow::stream::Location as StreamLocation;