+++ crates/sevenmark-parser/src/parser/macro/macro_newline.rs
@@ -7,5 +7,5 @@ use winnow::token::literal;
 pub fn macro_newline_parser(parser_input: &mut ParserInput) -> Result<SevenMarkElement> {
     literal("[br]").parse_next(parser_input)?;
 
-    Ok(SevenMarkElement::NewLine)
+    Ok(SevenMarkElement::HardBreak)
 }